text_to_ids function
First we need to turn the text into a number so the computer can
understand it. In the function text_to_ids(), we turn source_text and
target_text from words to ids. However, its important we also add the
<EOS> word id at the end of each sentence from target_text. This will
help the neural network predict when the sentence should end.